+++ spec/blazing/target_spec.rb
@@ -46,4 +46,22 @@ describe Blazing::Target do
     end
   end
 
+  describe '#rake_command' do
+
+    it 'prepends the environment variables specified in the rake call' do
+      config = Blazing::Config.new
+      config.rake :deploy, 'SOMEFUNKYVARIABLE=foobar'
+      target = Blazing::Target.new(:sometarget, '/path', config)
+      target.rake_command.should == 'SOMEFUNKYVARIABLE=foobar  bundle exec rake deploy'
+    end
+
+    it 'appends the RAILS_ENV specified as :rails_env option to the target call' do
+      config = Blazing::Config.new
+      config.rake :deploy
+      target = Blazing::Target.new(:sometarget, '/path', config, :rails_env => 'production')
+      target.rake_command.should == ' RAILS_ENV=production bundle exec rake deploy'
+    end
+
+  end
+
 end